Split clientapp_model into separate files and move enums
Built on current main. Splits Client, Vendor, and ServiceType into
their own model files. Moves VendorStatus and ComplianceStatus enums
into enums.py with the rest of the enums. Removes service_type field
from Vendor (dropped from ERD). Adds unique constraint to vendor_code
per updated ERD. Fixes missing well_bp import. Updates all imports.
