+++ vistaone-api/app/models/service_type.py
@@ -1,0 +1,19 @@
+from sqlalchemy.orm import mapped_column, relationship
+from sqlalchemy.sql import func
+import uuid
+from app.extensions import db
+
+
+class ServiceType(db.Model):
+    __tablename__ = "service_types"
+
+    service_type_id = mapped_column(
+        db.String(36), primary_key=True, default=lambda: str(uuid.uuid4())
+    )
+    service = mapped_column(db.String(255), nullable=False)
+    created_by = mapped_column(db.String(100))
+    created_date = mapped_column(db.DateTime, server_default=func.now())
+    last_modified_by = mapped_column(db.String(100))
+    last_modified_date = mapped_column(db.DateTime)
+
+    workorders = relationship("WorkOrder", back_populates="service_type")
